C++ Additional Files
Code is in C++ (for the parsing and creation of pathways inside the microarrays). All KEGG data documents were initially downloaded as .XML files but later converted to .txt files for ease of parsing.

UPDATE 2:

ProteinCleaner.cpp is a file that returns the gene names specifically for the Proteomics Core dataset of UTSW: it can and should be modified for files that are not of the same format.

UPDATE 1:

Parsing_KEGG_HSA+KO.cpp is a file that reads all the genes from a HSA or KO file (this can be easily modified to be any species however, all it requires is to rename the corresponding files accordingly). This quickly gathers all the genes in a pathway with all the corresponding names for each gene (I chose to create an algorithm that picks all data rather than miss maybe vital ones) and puts them into Pathways_HSA_X.txt files.

Cleaning_up_HSAEN_files.cpp is a file that removes duplicates in each pathway by finding only the unique values in each pathway.

Microarray_Analysis.cpp is a file that creates a hash table for each unique gene in the microarray. It then utilizes those hash values to then create hash values for each gene in every pathway, and then compares the names of the gene in the pathway to all genes in the hash table till it finds a match (can be altered to also see duplicates in the microarray by simply searching through the linked list of the corresponding hash number for all values of the gene as opposed to the first one).

UPDATE 0:

linkedlist.h <- header file in C++ that creates the construct linkedlist. I used several versions of it in each program, but the data structure is the same, just that I changed the key variable from a string to an int or double or pair depending on the file. Probably could just be kept as string by converting all int, double, and pairs into string values.

Parsing_Enzyme_KEGG.cpp is a file that reads through the Enzyme Bank of KEGG and exports a file called "Gene Bank.txt" wherein each enzyme's name is attached in a linked list. As randomly separating each enzyme is equivalent of a linear search, the enzymes for each hash number are then uploaded in the corresponding line of the text file. So the enzyme with hash number = 1 will belong to line 1, hash number = 3 belongs to line 3, and so on.

Parsing_KEGG_Data.cpp is a file that reads through "Gene Bank.txt" and through a corresponding KEGG file that is uploaded as an enzyme file, parsing through it to find the name of the pathway in KEGG as well as all the enzyme numbers to then create an enzyme list that contains all the enzymes per pathway and places them into Pathways_X.txt files. HOWEVER, the KEGG Enzyme Bank list is extremely messy with many labels given to an enzyme that is not displayed; as such, I only initially used these two files and updated it afterwards.